[chore] #249 - PointEditView의 enum과 class 파일로 분리
