+++ Carmu/Sources/Views/MyPage/CrewEdit/PointEdit/PointType.swift
@@ -5,4 +5,10 @@
 //  Created by 김영빈 on 2023/11/14.
 //
 
-import Foundation
+enum PointType {
+    case start
+    case stopover1
+    case stopover2
+    case stopover3
+    case end
+}
